+++ programm/object/visibleObject/shapes/plane/plane.h
@@ -15,7 +15,7 @@ class Plane : public BaseShape
 
 	}
 	void transform(const TransformParams& transformParams) override;
-	double intersection(const Ray& ray) override;
+	float intersection(const Ray& ray) override;
 	void setSpectralParams(float k_a,float k_d,float k_s);
 	void setColorParams(const ColorRGB& color);
 	Material getMaterial() override;
