+++ programm/object/visibleObject/shapes/sphere/sphere.h
@@ -19,7 +19,7 @@ class Sphere : public BaseShape
 	Sphere() = default;
 	Sphere(const VecD3& center,double radius,const Material& material);
 	void transform(const TransformParams& transformParams) override;
-	double intersection(const Ray& ray) override;
+	float intersection(const Ray& ray) override;
 	void setSpectralParams(float k_a,float k_d,float k_s);
 	void setColorParams(const ColorRGB& color);
 	Material getMaterial() override;
